Made `app` a global, not context value
The main difference between context and global values is that
context values cannot be accessed from imported macros by default:

{% from "helpers.html" import my_macro %} -
such macro is cached and thus cannot access current context;
for it to access context, it should be called like this -
{% from "helpers.html" import my_macro with context %}
which is less efficient.

`app` object is safe to use as global because its direct value will not change.

More info:
http://jinja.pocoo.org/docs/dev/api/#global-namespace